Flower UI tweaks

+++ src/main/java/edu/wpi/cs3733/d20/teamA/controllers/FlowerAdminController.java
@@ -176,11 +176,36 @@ public class FlowerAdminController extends AbstractController {
         new FlowerDialogController());
   }
 
+  private boolean hasDependentOrder(Flower flower) {
+    boolean constrained = false;
+    try {
+      for (Order order : flDatabase.orderOl()) {
+        if ((order.getFlowerType().equals(flower.getTypeFlower()))
+            && (order.getFlowerColor().equals(flower.getColor()))) {
+          constrained = true;
+        }
+      }
+    } catch (Exception e) {
+      DialogUtil.simpleErrorDialog(
+          dialogStackPane,
+          "Database Failure",
+          "Failed to verify that there were no outstanding orders for flower: "
+              + flower.toString());
+    }
+
+    return constrained;
+  }
+
   public void editFlower() {
     TreeItem<Flower> selection = tblFlowerView.getSelectionModel().getSelectedItem();
     if (selection != null) {
       Flower flower = selection.getValue();
-      FlowerDialogController controller = new FlowerDialogController(flower);
+
+      // Figure out whether any outstanding orders depend on this flower type, in which case we
+      // can't change the name / type
+      boolean constrained = hasDependentOrder(flower);
+
+      FlowerDialogController controller = new FlowerDialogController(flower, constrained);
       DialogUtil.complexDialog(
           dialogStackPane,
           "Edit Flower",
@@ -200,25 +225,26 @@ public class FlowerAdminController extends AbstractController {
     TreeItem<Flower> selected = tblFlowerView.getSelectionModel().getSelectedItem();
     if (selected != null) {
       Flower f = selected.getValue();
-      String name = f.getTypeFlower();
-      String color = f.getColor();
+      if (!hasDependentOrder(f)) {
+        String name = f.getTypeFlower();
+        String color = f.getColor();
 
-      try {
-        super.flDatabase.deleteFlower(name, color);
-      } catch (Exception e) {
-        e.printStackTrace();
-        DialogUtil.simpleErrorDialog(
-            dialogStackPane, "Error Deleting Flower", "Could not delete flower: " + f);
+        try {
+          super.flDatabase.deleteFlower(name, color);
+        } catch (Exception e) {
+          e.printStackTrace();
+          DialogUtil.simpleErrorDialog(
+              dialogStackPane, "Error Deleting Flower", "Could not delete flower: " + f);
+        }
+
+        update();
       }
-
-      update();
     } else {
       DialogUtil.simpleInfoDialog(
           dialogStackPane,
           "No Flower Selected",
           "Please select a flower by clicking a row in the table");
     }
-    update();
   }
 
   public void update() {
